gui: removed annoying useless 'TextInEditor' class
